requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.31
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.27.0
